+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.9.2] 2020
+### Improvements
+- Backup
+    - new **type** field on settings to specify the Timeshift backup mode: **RSYNC** or **BTRFS**
 ### UI
     - icons, buttons and colors changes